+++ wow_world_messages/src/world/vanilla/smsg_cast_result.rs
@@ -1241,9 +1241,7 @@ pub enum SMSG_CAST_RESULT_SimpleSpellCastResult {
 impl Default for SMSG_CAST_RESULT_SimpleSpellCastResult {
     fn default() -> Self {
         // First enumerator without any fields
-        Self::Success {
-            reason: Default::default(),
-        }
+        Self::Failure
     }
 }